requirements.txt
aliyun-python-sdk-core==2.4.4
aliyun-python-sdk-dysmsapi===1.0.0-SANPSHOT
certifi==2018.8.13
chardet==3.0.4
Django==1.11
django-cors-headers==2.4.0
django-qiniu-storage==2.3.1
djangorestframework==3.8.2
idna==2.7
Pillow==5.2.0
pycrypto==2.6.1
pytz==2018.5
qiniu==7.2.2
requests==2.19.1
six==1.11.0
urllib3==1.23
